+++ src/main/java/com/hx/blog_v2/controller/admin/BlogTagController.java
@@ -1,8 +1,14 @@
 package com.hx.blog_v2.controller.admin;
 
+import com.hx.blog_v2.domain.ErrorCode;
+import com.hx.blog_v2.domain.form.BeanIdForm;
 import com.hx.blog_v2.domain.form.BlogTagSaveForm;
+import com.hx.blog_v2.domain.validator.BeanIdValidator;
+import com.hx.blog_v2.domain.validator.BlogTagSaveValidator;
 import com.hx.blog_v2.service.interf.BlogTagService;
 import com.hx.common.interf.common.Result;
+import com.hx.common.util.ResultUtils;
+import com.hx.log.util.Tools;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -21,29 +27,48 @@ public class BlogTagController {
 
     @Autowired
     private BlogTagService blogTagService;
-
+    @Autowired
+    private BlogTagSaveValidator blogTagSaveValidator;
+    @Autowired
+    private BeanIdValidator beanIdValidator;
 
     @RequestMapping(value = "/add", method = RequestMethod.POST)
     public Result add(BlogTagSaveForm params) {
-
+        Result errResult = blogTagSaveValidator.validate(params, null);
+        if (!errResult.isSuccess()) {
+            return errResult;
+        }
+        if (!Tools.isEmpty(params.getId())) {
+            return ResultUtils.failed(ErrorCode.INPUT_NOT_FORMAT, " id 不为空 ! ");
+        }
 
         return blogTagService.add(params);
     }
 
     @RequestMapping(value = "/list", method = RequestMethod.GET)
     public Result list() {
-
         return blogTagService.list();
     }
 
     @RequestMapping(value = "/update", method = RequestMethod.POST)
     public Result update(BlogTagSaveForm params) {
+        Result errResult = blogTagSaveValidator.validate(params, null);
+        if (!errResult.isSuccess()) {
+            return errResult;
+        }
+        if (Tools.isEmpty(params.getId())) {
+            return ResultUtils.failed(ErrorCode.INPUT_NOT_FORMAT, " id 为空 ! ");
+        }
 
         return blogTagService.update(params);
     }
 
     @RequestMapping(value = "/remove", method = RequestMethod.POST)
-    public Result remove(BlogTagSaveForm params) {
+    public Result remove(BeanIdForm params) {
+        Result errResult = beanIdValidator.validate(params, null);
+        if (!errResult.isSuccess()) {
+            return errResult;
+        }
 
         return blogTagService.remove(params);
     }
