+++ src/main/java/com/hx/blog_v2/controller/admin/LinkController.java
@@ -1,11 +1,16 @@
 package com.hx.blog_v2.controller.admin;
 
+import com.hx.blog_v2.domain.ErrorCode;
 import com.hx.blog_v2.domain.form.BeanIdForm;
 import com.hx.blog_v2.domain.form.LinkSaveForm;
-import com.hx.blog_v2.domain.vo.AdminLinkVO;
+import com.hx.blog_v2.domain.validator.BeanIdValidator;
+import com.hx.blog_v2.domain.validator.LinkSaveValidator;
+import com.hx.blog_v2.domain.validator.PageValidator;
+import com.hx.blog_v2.domain.validator.ResourceInterfUpdateValidator;
 import com.hx.blog_v2.service.interf.LinkService;
 import com.hx.common.interf.common.Result;
-import com.hx.common.result.SimplePage;
+import com.hx.common.util.ResultUtils;
+import com.hx.log.util.Tools;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -24,9 +29,20 @@ public class LinkController {
 
     @Autowired
     private LinkService linkService;
+    @Autowired
+    private LinkSaveValidator linkSaveValidator;
+    @Autowired
+    private BeanIdValidator beanIdValidator;
 
     @RequestMapping(value = "/add", method = RequestMethod.POST)
     public Result add(LinkSaveForm params) {
+        Result errResult = linkSaveValidator.validate(params, null);
+        if (!errResult.isSuccess()) {
+            return errResult;
+        }
+        if (!Tools.isEmpty(params.getId())) {
+            return ResultUtils.failed(ErrorCode.INPUT_NOT_FORMAT, " id 不为空 ! ");
+        }
 
         return linkService.add(params);
     }
@@ -34,18 +50,28 @@ public class LinkController {
 
     @RequestMapping(value = "/list", method = RequestMethod.GET)
     public Result list() {
-
         return linkService.adminList();
     }
 
     @RequestMapping(value = "/update", method = RequestMethod.POST)
     public Result update(LinkSaveForm params) {
+        Result errResult = linkSaveValidator.validate(params, null);
+        if (!errResult.isSuccess()) {
+            return errResult;
+        }
+        if (Tools.isEmpty(params.getId())) {
+            return ResultUtils.failed(ErrorCode.INPUT_NOT_FORMAT, " id 为空 ! ");
+        }
 
         return linkService.update(params);
     }
 
     @RequestMapping(value = "/remove", method = RequestMethod.POST)
     public Result remove(BeanIdForm params) {
+        Result errResult = beanIdValidator.validate(params, null);
+        if (!errResult.isSuccess()) {
+            return errResult;
+        }
 
         return linkService.remove(params);
     }
